Added JWT token
Now client just needs to save the token and we have a rudimentary signup/in procedure. Shouldnt be too hard to add oAuth.
Beware the password is currently sent in plaintext however it is saved in db as a hash. We need a hash system on the client side to be complete and prevent mitm